Warn from getErrorReporter instead of throwing exception
